refactor: move repeated func def to utils folder
